requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.2.1
beautifulsoup4==4.14.2
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
filelock==3.13.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
google_search_results==2.4.2
groq==0.26.0
grpcio==1.74.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
MouseInfo==0.1.3
mpmath==1.3.0
msgspec==0.19.0
namex==0.1.0
narwhals==1.41.0
networkx==3.3
numpy==2.1.2
openai==1.61.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==5.29.5
pyarrow==20.0.0
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
pymongo==4.11.1
PyMsgBox==2.0.1
pyparsing==3.2.3
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.1.0
rpds-py==0.25.1
safetensors==0.5.2
scikit-image==0.25.2
scipy==1.16.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tifffile==2025.6.11
tokenizers==0.21.0
toml==0.10.2
torch==2.4.1
torch-directml==0.2.5.dev240914
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
